Proje ilk kurulum: Backend, Frontend ve Veritabanı yapısı tamamlandı

+++ src/main/java/com/kutuphane/AkilliKutuphane/repository/KategoriRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KategoriRepository extends JpaRepository<Kategori, Integer> {
+public interface KategoriRepository extends JpaRepository<Kategori, Long> {
 }
