+++ src/main/java/com/kutuphane/AkilliKutuphane/repository/YazarRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface YazarRepository extends JpaRepository<Yazar, Integer> {
+public interface YazarRepository extends JpaRepository<Yazar, Long> {
 }
